WIP: mgr/cephadm: Move all background methods into serve.py
I.e. move all expensive background things into serve in order
to make it harder to use them accidentially.
